etl/requirements.txt
python-dotenv==0.20.0
elasticsearch==7.7.0
elasticsearch-dsl==7.7.0
urllib3==1.26.8
pydantic=1.9.0
